+++ utils/logger.py
@@ -1,0 +1,14 @@
+from enum import StrEnum
+
+
+# Create a EMOJI enums for differnt types logger
+class LogLevelEmoji(StrEnum):
+    INFO = "ℹ️"
+    WARNING = "⚠️"
+    ERROR = "❌"
+    CRITICAL = "🚨"
+    DEBUG = "🐛"
+    COMPLETED = "✅"
+
+
+dividerString = "----------------------------------------------------------------------------------------------------------"
